+++ app/Http/Controllers/Api/EventController.php
@@ -7,38 +7,49 @@ use App\Http\Resources\EventResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
     use CanLoadRelationships;
 
-    private  $relations = ['user', 'attendees', 'attendees.user'];
+    private array $relations = ['user', 'attendees', 'attendees.user'];
+
+    // public function __construct()
+    // {
+    //     $this->middleware('auth:sanctum')->except(['index', 'show']);
+    //     $this->middleware('throttle:api')
+    //         ->only(['store', 'update', 'destroy']);
+    //     $this->authorizeResource(Event::class, 'event');
+    // }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        Gate::authorize('viewAny', Event::class);
         $query = $this->loadRelationships(Event::query());
 
-
-        return EventResource::collection($query->latest()->paginate());
+        return EventResource::collection(
+            $query->latest()->paginate()
+        );
     }
-
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
-
+        Gate::authorize('create', Event::class);
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
                 'description' => 'nullable|string',
                 'start_time' => 'required|date',
-                'end_time' => 'required|date|after:start_time',
+                'end_time' => 'required|date|after:start_time'
             ]),
-            'user_id' => 1
+            'user_id' => $request->user()->id
         ]);
 
         return new EventResource($this->loadRelationships($event));
@@ -49,7 +60,10 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
-        return new EventResource($this->loadRelationships($event));
+        Gate::authorize('view', $event);
+        return new EventResource(
+            $this->loadRelationships($event)
+        );
     }
 
     /**
@@ -57,12 +71,17 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+        // if (Gate::denies('update-event', $event)) {
+        //     abort(403, 'You are not authorized to update this event.');
+        // }
+        // $this->authorize('update-event', $event);
+        Gate::authorize('update', $event);
         $event->update(
             $request->validate([
                 'name' => 'sometimes|string|max:255',
                 'description' => 'nullable|string',
                 'start_time' => 'sometimes|date',
-                'end_time' => 'sometimes|date|after:start_time',
+                'end_time' => 'sometimes|date|after:start_time'
             ])
         );
 
@@ -74,10 +93,9 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        Gate::authorize('delete', $event);
         $event->delete();
 
-        return response()->json([
-            'message' => 'Event deleted'
-        ]);
+        return response(status: 204);
     }
 }
